# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Music Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface Guild {
  id: string;
  name: string;
  icon?: string | null;
  memberCount: number;
  hasActivePlayer: boolean;
}

export interface Track {
  encoded: string;
  identifier: string;
  title: string;
  author: string;
  duration: number;
  thumbnail?: string | null;
  uri?: string | null;
  sourceName: string;
  isStream: boolean;
  requestedBy?: string | null;
}

export type PlayerStateLoopMode =
  (typeof PlayerStateLoopMode)[keyof typeof PlayerStateLoopMode];

export const PlayerStateLoopMode = {
  none: "none",
  track: "track",
  queue: "queue",
} as const;

export interface PlayerState {
  guildId: string;
  isPlaying: boolean;
  isPaused: boolean;
  volume: number;
  position: number;
  loopMode: PlayerStateLoopMode;
  currentTrack?: Track | null;
  voiceChannelId?: string | null;
  voiceChannelName?: string | null;
  queueSize: number;
  nodeConnected: boolean;
}

export interface QueueResponse {
  guildId: string;
  tracks: Track[];
  totalDuration: number;
}

export type PlayRequestSource =
  (typeof PlayRequestSource)[keyof typeof PlayRequestSource];

export const PlayRequestSource = {
  youtube: "youtube",
  soundcloud: "soundcloud",
  spotify: "spotify",
  ytsearch: "ytsearch",
  scsearch: "scsearch",
} as const;

export interface PlayRequest {
  query: string;
  source?: PlayRequestSource;
  voiceChannelId?: string;
}

export interface PlayResponse {
  success: boolean;
  message: string;
  track?: Track | null;
  playlistName?: string | null;
  tracksAdded: number;
}

export interface VolumeRequest {
  /**
   * @minimum 0
   * @maximum 200
   */
  volume: number;
}

export interface SeekRequest {
  position: number;
}

export type LoopRequestMode =
  (typeof LoopRequestMode)[keyof typeof LoopRequestMode];

export const LoopRequestMode = {
  none: "none",
  track: "track",
  queue: "queue",
} as const;

export interface LoopRequest {
  mode: LoopRequestMode;
}

export interface SearchResponse {
  loadType: string;
  tracks: Track[];
  playlistName?: string | null;
}

export interface GuildSettings {
  guildId: string;
  prefix: string;
  defaultVolume: number;
  djRoleId?: string | null;
  textChannelId?: string | null;
  autoplay: boolean;
  announceNowPlaying: boolean;
}

export interface GuildSettingsUpdate {
  prefix?: string;
  defaultVolume?: number;
  djRoleId?: string | null;
  textChannelId?: string | null;
  autoplay?: boolean;
  announceNowPlaying?: boolean;
}

export interface BotStats {
  guildCount: number;
  activePlayers: number;
  totalTracksPlayed: number;
  uptimeMs: number;
  nodeConnected: boolean;
  memoryUsageMb: number;
}

export interface HistoryEntry {
  id: number;
  guildId: string;
  guildName?: string | null;
  title: string;
  author: string;
  duration: number;
  thumbnail?: string | null;
  uri?: string | null;
  sourceName: string;
  playedAt: string;
  requestedBy?: string | null;
}

export interface NodeStatus {
  connected: boolean;
  host: string;
  port: number;
  players: number;
  playingPlayers: number;
  uptime?: number;
  memoryFree?: number;
  memoryUsed?: number;
  cpuCores?: number;
  cpuLoad?: number;
}

export type SearchTracksParams = {
  q: string;
  source?: SearchTracksSource;
};

export type SearchTracksSource =
  (typeof SearchTracksSource)[keyof typeof SearchTracksSource];

export const SearchTracksSource = {
  youtube: "youtube",
  soundcloud: "soundcloud",
  spotify: "spotify",
  ytsearch: "ytsearch",
  scsearch: "scsearch",
} as const;

export type GetPlayHistoryParams = {
  limit?: number;
};