# lib/api-zod/src/generated/types/botStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Music Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface BotStats {
  guildCount: number;
  activePlayers: number;
  totalTracksPlayed: number;
  uptimeMs: number;
  nodeConnected: boolean;
  memoryUsageMb: number;
}