# lib/api-zod/src/generated/types/historyEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Music Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryEntry {
  id: number;
  guildId: string;
  guildName?: string | null;
  title: string;
  author: string;
  duration: number;
  thumbnail?: string | null;
  uri?: string | null;
  sourceName: string;
  playedAt: Date;
  requestedBy?: string | null;
}