# lib/api-zod/src/generated/types/loopRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Music Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { LoopRequestMode } from "./loopRequestMode";

export interface LoopRequest {
  mode: LoopRequestMode;
}