# lib/api-zod/src/generated/types/playResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Music Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { Track } from "./track";

export interface PlayResponse {
  success: boolean;
  message: string;
  track?: Track | null;
  playlistName?: string | null;
  tracksAdded: number;
}