# lib/api-zod/src/generated/types/track.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Music Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Track {
  encoded: string;
  identifier: string;
  title: string;
  author: string;
  duration: number;
  thumbnail?: string | null;
  uri?: string | null;
  sourceName: string;
  isStream: boolean;
  requestedBy?: string | null;
}