# lib/api-zod/src/generated/types/volumeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Music Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface VolumeRequest {
  /**
   * @minimum 0
   * @maximum 200
   */
  volume: number;
}